Fix map output
