+++ cosmos_governance_parameters/proto/events.proto
@@ -1,0 +1,35 @@
+syntax = "proto3";
+
+package cosmos.custom_events;
+
+message ProposalEvents {
+    repeated GovParamsOptional gov_params_changes = 1;
+    repeated string passed_proposal_ids = 2;
+}
+
+message DepositParamsOptional {
+    repeated Deposit min_deposit = 1;
+    optional string max_deposit_period = 2;
+}
+
+message Deposit {
+    string denom = 1;
+    string amount = 2;
+}
+
+message VotingParamsOptional {
+    optional string voting_period = 1;
+}
+
+message TallyParamsOptional {
+    optional string quorum = 1;
+    optional string threshold = 2;
+    optional string veto_threshold = 3;
+}
+
+message GovParamsOptional {
+    string proposal_id = 1;
+    optional DepositParamsOptional deposit_params = 2;
+    optional VotingParamsOptional voting_params = 3;
+    optional TallyParamsOptional tally_params = 4;
+}
