+++ cosmos_governance_parameters/src/gov_params.rs
@@ -7,8 +7,8 @@ use cosmos_proposals::genesis_params::GovParams;
 use substreams::store::StoreSet;
 use substreams::store::StoreSetString;
 
-use crate::pb::pb::events::GovParamsOptional;
-use crate::pb::pb::events::ProposalEvents;
+use crate::pb::cosmos::custom_events::ProposalEvents;
+use crate::serde_structs::GovParamsOptional as GovParamsSerializable;
 
 #[substreams::handlers::store]
 pub fn gov_params(
@@ -53,7 +53,8 @@ fn set_genesis_params(gov_params: &StoreSetString, genesis_params: &String) {
 }
 
 fn set_new_gov_params(gov_params: &StoreSetString, gov_param_proposal_str: &String) {
-    let parsed: GovParamsOptional = serde_json::from_str(&gov_param_proposal_str).expect("Failed to parse gov params");
+    let parsed: GovParamsSerializable =
+        serde_json::from_str(&gov_param_proposal_str).expect("Failed to parse gov params");
 
     if let Some(deposit_params) = parsed.deposit_params {
         if deposit_params.min_deposit.len() > 0 {
