+++ cosmos_governance_parameters/src/lib.rs
@@ -2,3 +2,4 @@ mod gov_params;
 mod map_events;
 mod pb;
 mod pending_gov_params;
+mod serde_structs;
