+++ cosmos_governance_parameters/src/pb/mod.rs
@@ -1,5 +1,8 @@
-pub mod pb {
-    pub mod events {
-        include!("events.rs");
+// @generated
+pub mod cosmos {
+    // @@protoc_insertion_point(attribute:cosmos.custom_events)
+    pub mod custom_events {
+        include!("cosmos.custom_events.rs");
+        // @@protoc_insertion_point(cosmos.custom_events)
     }
 }
